Adding todo and some more int to NSInteger changes, working on Cocoa currently

+++ StreamLineReader.h
@@ -38,7 +38,7 @@
 
 - (BOOL) getMoreCharacters;
 - (NSString *) extractNextLine;
-- (int) delimPosInBuffer;
+- (NSUInteger) delimPosInBuffer;
 - (BOOL) canExtractNextLine;
 
 @end
